fix: use relative import and extract RateInfoService types
