+++ src/lib/nodes/rate-info-service/types/RateHistoryInfoResponse.ts
@@ -1,0 +1,9 @@
+export type RateHistoryInfoResponse = {
+  success: boolean
+  date: number
+  result: {
+    _id: string
+    date: number
+    tickers: Record<string, number>
+  }[]
+}
