+++ src/lib/nodes/rate-info-service/types/RateInfoResponse.ts
@@ -1,0 +1,5 @@
+export type RateInfoResponse = {
+  success: boolean
+  date: number
+  result: Record<string, number>
+}
