+++ src/store/modules/rate/index.ts
@@ -1,10 +1,10 @@
 import { Module } from 'vuex'
 import { RootState } from '@/store/types'
-import { RateState } from '@/store/modules/rate/types.ts'
-import { state } from '@/store/modules/rate/rate-state.ts'
-import { mutations } from '@/store/modules/rate/rate-mutations.ts'
-import { actions } from '@/store/modules/rate/rate-actions.ts'
-import { getters } from '@/store/modules/rate/rate-getters.ts'
+import { RateState } from '@/store/modules/rate/types'
+import { state } from '@/store/modules/rate/rate-state'
+import { mutations } from '@/store/modules/rate/rate-mutations'
+import { actions } from '@/store/modules/rate/rate-actions'
+import { getters } from '@/store/modules/rate/rate-getters'
 
 const rateModule: Module<RateState, RootState> = {
   namespaced: true,
