+++ src/store/modules/rate/rate-getters.ts
@@ -1,6 +1,6 @@
 import { GetterTree } from 'vuex'
 import { RootState } from '@/store/types'
-import { RateState } from '@/store/modules/rate/types.ts'
+import { RateState } from '@/store/modules/rate/types'
 
 export const getters: GetterTree<RateState, RootState> = {
   historyRate:
