+++ src/store/modules/rate/rate-state.ts
@@ -1,4 +1,4 @@
-import { RateState } from '@/store/modules/rate/types.ts'
+import { RateState } from '@/store/modules/rate/types'
 
 export const state: RateState = {
   rates: {},
