+++ src/store/modules/services/index.ts
@@ -1,10 +1,10 @@
-import { state } from './services-state.ts'
-import { actions } from './services-actions.ts'
-import { mutations } from './services-mutations.ts'
-import { getters } from './services-getters.ts'
+import { state } from './services-state'
+import { actions } from './services-actions'
+import { mutations } from './services-mutations'
+import { getters } from './services-getters'
 import { Module } from 'vuex'
 import { RootState } from '@/store/types'
-import { ServicesState } from '@/store/modules/services/types.ts'
+import { ServicesState } from '@/store/modules/services/types'
 
 const servicesModule: Module<ServicesState, RootState> = {
   namespaced: true,
