+++ src/store/modules/services/services-actions.ts
@@ -1,7 +1,7 @@
 import { services } from '@/lib/nodes/services'
 import { ActionTree } from 'vuex'
 import { RootState } from '@/store/types'
-import { ServicesState } from '@/store/modules/services/types.ts'
+import { ServicesState } from '@/store/modules/services/types'
 
 export const actions: ActionTree<ServicesState, RootState> = {
   updateStatus() {
